requirements.txt
streamlit~=1.21.0
requests~=2.28.2
fer~=22.5.0
matplotlib~=3.7.1
Pillow~=9.5.0